Move more headers from mapping/ to internal. (#1796)
The moved headers are not used in any public API.
