+++ cartographer/mapping/imu_tracker_test.cc
@@ -17,7 +17,7 @@
 #include "cartographer/mapping/imu_tracker.h"
 
 #include "absl/memory/memory.h"
-#include "cartographer/mapping/eigen_quaterniond_from_two_vectors.h"
+#include "cartographer/mapping/internal/eigen_quaterniond_from_two_vectors.h"
 #include "gtest/gtest.h"
 
 namespace cartographer {
