+++ cartographer/mapping/internal/2d/scan_matching/tsdf_match_cost_function_2d_test.cc
@@ -19,7 +19,7 @@
 #include "cartographer/common/lua_parameter_dictionary.h"
 #include "cartographer/common/lua_parameter_dictionary_test_helpers.h"
 #include "cartographer/mapping/2d/tsdf_2d.h"
-#include "cartographer/mapping/2d/tsdf_range_data_inserter_2d.h"
+#include "cartographer/mapping/internal/2d/tsdf_range_data_inserter_2d.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
