+++ cartographer/mapping/internal/2d/tsdf_range_data_inserter_2d.h
