+++ cartographer/mapping/internal/eigen_quaterniond_from_two_vectors.h
