+++ cartographer/mapping/pose_extrapolator_interface.cc
@@ -18,7 +18,7 @@
 
 #include "cartographer/common/ceres_solver_options.h"
 #include "cartographer/common/time.h"
-#include "cartographer/mapping/imu_based_pose_extrapolator.h"
+#include "cartographer/mapping/internal/imu_based_pose_extrapolator.h"
 #include "cartographer/mapping/pose_extrapolator.h"
 
 namespace cartographer {
